+++ src/Tetrium.cpp
@@ -20,8 +20,6 @@
 #include "imgui.h"
 #include "lib/ImGuiUtils.h"
 
-#include "lib/Utils.h"
-
 // Molten VK Config
 #if __APPLE__
 #include "MoltenVKConfig.h"
@@ -582,47 +580,6 @@ void Tetrium::Init(const Tetrium::InitOptions& options)
     createFunnyObjects();
 }
 
-void Tetrium::Run()
-{
-    ASSERT(_window);
-    glfwShowWindow(_window);
-    while (!glfwWindowShouldClose(_window)) {
-        glfwPollEvents();
-        Tick();
-    }
-}
-
-void Tetrium::Tick()
-{
-    if (_paused) {
-        std::this_thread::yield();
-        return;
-    }
-    _deltaTimer.Tick();
-    {
-        PROFILE_SCOPE(&_profiler, "Main Tick");
-        {
-            PROFILE_SCOPE(&_profiler, "CPU");
-            // CPU-exclusive workloads
-            double deltaTime = _deltaTimer.GetDeltaTime();
-            _timeSinceStartSeconds += deltaTime;
-            _timeSinceStartNanoSeconds += _deltaTimer.GetDeltaTimeNanoSeconds();
-            _inputManager.Tick(deltaTime);
-            TickContext tickData{&_mainCamera, deltaTime};
-            tickData.profiler = &_profiler;
-            flushEngineUBOStatic(_currentFrame);
-            drawFrame(&tickData, _currentFrame);
-            _currentFrame = (_currentFrame + 1) % NUM_FRAME_IN_FLIGHT;
-        }
-        {
-            PROFILE_SCOPE(&_profiler, "GPU");
-            vkDeviceWaitIdle(this->_device->logicalDevice);
-        }
-    }
-    _lastProfilerData = _profiler.NewProfile();
-    _numTicks++;
-}
-
 void Tetrium::framebufferResizeCallback(GLFWwindow* window, int width, int height)
 {
     DEBUG("Window resized to {}x{}", width, height);
@@ -651,110 +608,6 @@ void Tetrium::createDevice()
     this->_deletionStack.push([this]() { this->_device->Cleanup(); });
 }
 
-void Tetrium::setupSoftwareEvenOddFrame()
-{
-    DEBUG("Setting up even-odd frame resources...");
-    auto& ctx = _softwareEvenOddCtx;
-    ctx.timeEngineStart = std::chrono::steady_clock::now();
-
-    // get refresh cycle
-    VkRefreshCycleDurationGOOGLE refreshCycleDuration;
-    ASSERT(_device->logicalDevice);
-
-    PFN_vkGetRefreshCycleDurationGOOGLE ptr = reinterpret_cast<PFN_vkGetRefreshCycleDurationGOOGLE>(
-        vkGetInstanceProcAddr(_instance, "vkGetRefreshCycleDurationGOOGLE")
-    );
-    ASSERT(ptr);
-    VK_CHECK_RESULT(ptr(_device->logicalDevice, _swapChain.chain, &refreshCycleDuration));
-    ctx.nanoSecondsPerFrame = refreshCycleDuration.refreshDuration;
-    ASSERT(ctx.nanoSecondsPerFrame != 0);
-}
-
-void Tetrium::checkSoftwareEvenOddFrameSupport() {}
-
-void Tetrium::setupHardwareEvenOddFrame()
-{
-#if WIN32
-    NEEDS_IMPLEMENTATION();
-#endif
-#if __APPLE__
-    NEEDS_IMPLEMENTATION();
-#endif
-    DEBUG("Setting up even-odd frame resources...");
-    auto& ctx = _hardWareEvenOddCtx;
-    ctx.vkGetSwapchainCounterEXT = reinterpret_cast<PFN_vkGetSwapchainCounterEXT>(
-        vkGetInstanceProcAddr(_instance, "vkGetSwapchainCounterEXT")
-    );
-    if (ctx.vkGetSwapchainCounterEXT == nullptr) {
-        PANIC("Failed to get function pointer to {}", "vkGetSwapchainCounterEXT");
-    }
-}
-
-// check for hardware and software support for even-odd frame rendering.
-// specifically, the GPU must provide a surface counter support --
-// the counter ticks every time a vertical blanking period occurs,
-// which we use to decide whether the next frame to present should be
-// even or odd.
-void Tetrium::checkHardwareEvenOddFrameSupport()
-{
-#ifdef WIN32
-    NEEDS_IMPLEMENTATION();
-#endif // WIN32
-#ifndef __linux__
-    return;
-#endif // __linux__
-    // check instance extensions
-    uint32_t numExtensions = 0;
-    VkResult result = VK_SUCCESS;
-
-    VK_CHECK_RESULT(vkEnumerateInstanceExtensionProperties(nullptr, &numExtensions, nullptr));
-    std::vector<VkExtensionProperties> extensions(numExtensions);
-    VK_CHECK_RESULT(
-        vkEnumerateInstanceExtensionProperties(nullptr, &numExtensions, extensions.data())
-    );
-
-    std::unordered_set<std::string> evenOddExtensions(
-        DEFAULTS::Engine::EVEN_ODD_HARDWARE_INSTANCE_EXTENSIONS.begin(),
-        DEFAULTS::Engine::EVEN_ODD_HARDWARE_INSTANCE_EXTENSIONS.end()
-    );
-
-    for (VkExtensionProperties& property : extensions) {
-        evenOddExtensions.erase(std::string(property.extensionName));
-    }
-
-    if (!evenOddExtensions.empty()) {
-        ERROR("The following instance extensions required for even-odd rendering is not available: "
-        );
-        for (const std::string& extension : evenOddExtensions) {
-            ERROR(extension);
-        }
-        PANIC("Even-odd frame not supported!");
-    }
-
-    VkSurfaceCapabilities2EXT capabilities{
-        .sType = VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_EXT, .pNext = VK_NULL_HANDLE
-    };
-
-    auto func = (PFN_vkGetPhysicalDeviceSurfaceCapabilities2EXT
-    )vkGetInstanceProcAddr(_instance, "vkGetPhysicalDeviceSurfaceCapabilities2EXT");
-    if (!func) {
-        PANIC(
-            "Failed to find function pointer to {}", "vkGetPhysicalDeviceSurfaceCapabilities2EXT"
-        );
-    }
-    ASSERT(_mainProjectorDisplay.surface && _mainProjectorDisplay.display);
-    VK_CHECK_RESULT(func(_device->physicalDevice, _mainProjectorDisplay.surface, &capabilities));
-
-    bool hasVerticalBlankingCounter
-        = (capabilities.supportedSurfaceCounters
-           & VkSurfaceCounterFlagBitsEXT::VK_SURFACE_COUNTER_VBLANK_BIT_EXT)
-          != 0;
-    if (!hasVerticalBlankingCounter) {
-        PANIC("Even-odd frame not supported!");
-    }
-
-    DEBUG("Even-odd frame support check passed!");
-}
 
 void Tetrium::initVulkan()
 {
@@ -911,8 +764,7 @@ void Tetrium::createInstance()
     createInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
     createInfo.pApplicationInfo = &appInfo;
 
-    std::vector<const char*> instanceExtensions
-        = DEFAULTS::Engine::DEFAULT_INSTANCE_EXTENSIONS;
+    std::vector<const char*> instanceExtensions = DEFAULTS::Engine::DEFAULT_INSTANCE_EXTENSIONS;
     // get glfw Extensions
     {
         uint32_t glfwExtensionCount = 0;
@@ -985,19 +837,6 @@ void Tetrium::createInstance()
     _deletionStack.push([this]() { vkDestroyInstance(this->_instance, nullptr); });
 
     INFO("Vulkan instance created.");
-}
-
-void Tetrium::createGlfwWindowSurface()
-{
-    NEEDS_IMPLEMENTATION();
-    // VkResult result
-    //     = glfwCreateWindowSurface(this->_instance, this->_window, nullptr, &this->_surface);
-    // if (result != VK_SUCCESS) {
-    //     FATAL("Failed to create window surface.");
-    // }
-    // _deletionStack.push([this]() { vkDestroySurfaceKHR(this->_instance, this->_surface, nullptr);
-    // }
-    // );
 }
 
 void Tetrium::populateDebugMessengerCreateInfo(VkDebugUtilsMessengerCreateInfoEXT& createInfo)
@@ -1165,7 +1004,6 @@ void Tetrium::createSwapChain(Tetrium::SwapChainContext& ctx, const VkSurfaceKHR
 
     VkSwapchainCreateInfoKHR createInfo{};
     createInfo.sType = VK_STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR;
-    // createInfo.surface = _surface;
     createInfo.surface = surface;
 
     createInfo.minImageCount = imageCount;
@@ -1682,251 +1520,6 @@ void Tetrium::createDepthBuffer(SwapChainContext& ctx)
     );
 }
 
-void Tetrium::flushEngineUBOStatic(uint8_t frame)
-{
-    VQBuffer& buf = _engineUBOStatic[frame];
-    EngineUBOStatic ubo{
-        _mainCamera.GetViewMatrix() // view
-        // proj
-    };
-    getMainProjectionMatrix(ubo.proj);
-    ubo.timeSinceStartSeconds = _timeSinceStartSeconds;
-    ubo.sinWave = (sin(_timeSinceStartSeconds) + 1) / 2.f; // offset to [0, 1]
-    memcpy(buf.bufferAddress, &ubo, sizeof(ubo));
-}
-
-void Tetrium::drawFrame(TickContext* ctx, uint8_t frame)
-{
-    SyncPrimitives& sync = _syncProjector[frame];
-
-    //  Wait for the previous frame to finish
-    PROFILE_SCOPE(&_profiler, "Render Tick");
-    vkWaitForFences(_device->logicalDevice, 1, &sync.fenceInFlight, VK_TRUE, UINT64_MAX);
-
-    //  Acquire an image from the swap chain
-    uint32_t swapchainImageIndex;
-    VkResult result = vkAcquireNextImageKHR(
-        this->_device->logicalDevice,
-        _swapChain.chain,
-        UINT64_MAX,
-        sync.semaImageAvailable,
-        VK_NULL_HANDLE,
-        &swapchainImageIndex
-    );
-    [[unlikely]] if (result == VK_ERROR_OUT_OF_DATE_KHR || result == VK_SUBOPTIMAL_KHR) {
-        this->recreateSwapChain(_swapChain);
-        recreateVirtualFrameBuffers();
-        return;
-    } else [[unlikely]] if (result != VK_SUCCESS) {
-        const char* res = string_VkResult(result);
-        PANIC("Failed to acquire swap chain image: {}", res);
-    }
-
-    // lock the fence
-    vkResetFences(this->_device->logicalDevice, 1, &sync.fenceInFlight);
-
-    vk::CommandBuffer CB1(_device->graphicsCommandBuffers[frame]);
-    //  Record a command buffer which draws the scene onto that image
-    CB1.reset();
-    { // update ctx->graphics field
-        ctx->graphics.currentFrameInFlight = frame;
-        ctx->graphics.currentSwapchainImageIndex = swapchainImageIndex;
-        ctx->graphics.CB = CB1;
-        // ctx->graphics.currentFB = FB;
-        ctx->graphics.currentFBextend = _swapChain.extent;
-        getMainProjectionMatrix(ctx->graphics.mainProjectionMatrix);
-    }
-
-    { // begin command buffer
-        VkCommandBufferBeginInfo beginInfo{};
-        beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-        beginInfo.flags = 0;                  // Optional
-        beginInfo.pInheritanceInfo = nullptr; // Optional
-        CB1.begin(vk::CommandBufferBeginInfo());
-    }
-
-    { // main render pass
-        vk::Extent2D extend = _swapChain.extent;
-        vk::Rect2D renderArea(VkOffset2D{0, 0}, extend);
-        vk::RenderPassBeginInfo renderPassBeginInfo(
-            {}, {}, renderArea, _clearValues.size(), _clearValues.data(), nullptr
-        );
-
-        VkViewport viewport{};
-        viewport.x = 0.0f;
-        viewport.y = 0.0f;
-        viewport.width = static_cast<float>(extend.width);
-        viewport.height = static_cast<float>(extend.height);
-        viewport.minDepth = 0.0f;
-        viewport.maxDepth = 1.0f;
-
-        VkRect2D scissor{};
-        scissor.offset = {0, 0};
-        scissor.extent = extend;
-
-        // two-pass rendering: render RGB and OCV colors onto two virtual FBs
-        {
-            renderPassBeginInfo.renderPass = _renderContexts.RGB.renderPass;
-            renderPassBeginInfo.framebuffer
-                = _renderContexts.RGB.virtualFrameBuffer
-                      .frameBuffer[swapchainImageIndex]; // which frame buffer in the
-                                                         // swapchain do the pass i.e.
-                                                         // all draw calls render to?
-            CB1.beginRenderPass(renderPassBeginInfo, vk::SubpassContents::eInline);
-            vkCmdSetViewport(CB1, 0, 1, &viewport);
-            vkCmdSetScissor(CB1, 0, 1, &scissor);
-            _renderer.TickRGB(ctx);
-            CB1.endRenderPass();
-
-            drawImGui(ColorSpace::RGB);
-            _imguiManager.RecordCommandBufferRGB(CB1, extend, swapchainImageIndex);
-        }
-        {
-            renderPassBeginInfo.renderPass = _renderContexts.OCV.renderPass;
-            renderPassBeginInfo.framebuffer
-                = _renderContexts.OCV.virtualFrameBuffer
-                      .frameBuffer[swapchainImageIndex]; // which frame buffer in the
-                                                         // swapchain do the pass i.e.
-                                                         // all draw calls render to?
-            CB1.beginRenderPass(renderPassBeginInfo, vk::SubpassContents::eInline);
-            vkCmdSetViewport(CB1, 0, 1, &viewport);
-            vkCmdSetScissor(CB1, 0, 1, &scissor);
-            _renderer.TickOCV(ctx);
-            CB1.endRenderPass();
-
-            drawImGui(ColorSpace::OCV);
-            _imguiManager.RecordCommandBufferOCV(CB1, extend, swapchainImageIndex);
-        }
-    }
-
-    CB1.end();
-
-    VkSubmitInfo submitInfo{.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO};
-
-    VkSemaphore waitSemaphores[] = {}; // don't need semaphore since we're drawing on virtual fb
-    VkPipelineStageFlags waitStages[]
-        = {VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT}; // wait for color to be available
-    std::array<VkCommandBuffer, 1> submitCommandBuffers = {CB1};
-
-    submitInfo.waitSemaphoreCount = 0;
-    submitInfo.pWaitSemaphores = waitSemaphores;
-    submitInfo.pWaitDstStageMask = waitStages;
-    submitInfo.commandBufferCount = static_cast<uint32_t>(submitCommandBuffers.size());
-    submitInfo.pCommandBuffers = submitCommandBuffers.data();
-
-    VkSemaphore signalSemaphores[] = {sync.semaRenderFinished};
-    submitInfo.signalSemaphoreCount = 1;
-    submitInfo.pSignalSemaphores = signalSemaphores;
-
-    if (vkQueueSubmit(_device->graphicsQueue, 1, &submitInfo, VK_NULL_HANDLE) != VK_SUCCESS) {
-        FATAL("Failed to submit draw command buffer!");
-    }
-
-    vk::CommandBuffer CB2(_device->graphicsCommandBuffers2[frame]);
-    CB2.reset();
-    CB2.begin(vk::CommandBufferBeginInfo());
-
-    // choose whether to render the even/odd frame buffer, discarding the other
-    bool isEven = isEvenFrame();
-    VkImage virtualFramebufferImage
-        = isEven ? _renderContexts.RGB.virtualFrameBuffer.image[swapchainImageIndex]
-                 : _renderContexts.OCV.virtualFrameBuffer.image[swapchainImageIndex];
-
-    VkImage swapchainFramebufferImage = _swapChain.image[swapchainImageIndex];
-    Utils::ImageTransfer::CmdCopyToFB(
-        CB2, virtualFramebufferImage, swapchainFramebufferImage, _swapChain.extent
-    );
-
-    if (isEven != _evenOddDebugCtx.currShouldBeEven) {
-        _evenOddDebugCtx.numDroppedFrames++;
-    }
-    _evenOddDebugCtx.currShouldBeEven = !isEven; // advance to next frame
-
-    // virtual FB has been copied onto the physical FB, paint ImGui now.
-    ctx->graphics.CB = CB2;
-    CB2.end();
-
-    // submit CB2
-    VkSubmitInfo submitInfo2{.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO};
-    std::array<VkCommandBuffer, 1> submitCommandBuffers2 = {CB2};
-    // for the image transfer, two semas need to be uppsed:
-    // 1. the render from the previous CB has to finish for 2 virtual FBs to be available
-    // 2. the actual FB needs to be available for copying
-    VkSemaphore waitSemaphores2[] = {sync.semaImageAvailable, sync.semaRenderFinished};
-    VkPipelineStageFlags waitStages2[] // both sema need to pass before the image transfer happens
-        = {VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT, VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT};
-
-    submitInfo2.waitSemaphoreCount = 2;
-    submitInfo2.pWaitSemaphores = waitSemaphores2;
-    submitInfo2.pWaitDstStageMask = waitStages2;
-    submitInfo2.commandBufferCount = static_cast<uint32_t>(submitCommandBuffers2.size());
-    submitInfo2.pCommandBuffers = submitCommandBuffers2.data();
-
-    VkSemaphore signalSemaphores2[]
-        = {sync.semaImageCopyFinished}; // signal this semaphore
-                                        // for the presentation to happen
-    submitInfo2.signalSemaphoreCount = 1;
-    submitInfo2.pSignalSemaphores = signalSemaphores2;
-
-    if (vkQueueSubmit(_device->graphicsQueue, 1, &submitInfo2, sync.fenceInFlight) != VK_SUCCESS) {
-        FATAL("Failed to submit draw command buffer!");
-    }
-
-    //  Present the swap chain image
-    VkPresentInfoKHR presentInfo{};
-    presentInfo.sType = VK_STRUCTURE_TYPE_PRESENT_INFO_KHR;
-
-    // wait for sync.semaRenderFinished upped by the previous render command
-    presentInfo.waitSemaphoreCount = 1;
-    presentInfo.pWaitSemaphores = signalSemaphores2; // semaImageCopyFinished, virtual fb
-                                                     // is copied over to actual fb, can render
-
-    VkSwapchainKHR swapChains[] = {_swapChain.chain};
-
-    presentInfo.swapchainCount = 1;
-    presentInfo.pSwapchains = swapChains;
-    presentInfo.pImageIndices = &swapchainImageIndex; // specify which image to present
-    presentInfo.pResults = nullptr;                   // Optional: can be used to check if
-                                                      // presentation was successful
-                                                      //
-
-    uint64_t time = 0;
-#if VIRTUAL_VSYNC
-    if (_softwareEvenOddCtx.mostRecentPresentFinish) {
-        time = _softwareEvenOddCtx.mostRecentPresentFinish
-               + _softwareEvenOddCtx.nanoSecondsPerFrame * _softwareEvenOddCtx.vsyncFrameOffset;
-    }
-#endif // VIRTUAL_VSYNC
-
-    // label each frame with the tick number
-    // this is useful for calculating the virtual frame counter,
-    // as we can take the max(frame.id) to get the number of presented frames.
-
-    VkPresentTimeGOOGLE presentTime{(uint32_t)_numTicks, time};
-
-    VkPresentTimesInfoGOOGLE presentTimeInfo{
-        .sType = VK_STRUCTURE_TYPE_PRESENT_TIMES_INFO_GOOGLE,
-        .pNext = VK_NULL_HANDLE,
-        .swapchainCount = 1,
-        .pTimes = &presentTime
-    };
-
-    if (_tetraMode == TetraMode::kEvenOddSoftwareSync) {
-        presentInfo.pNext = &presentTimeInfo;
-    }
-
-    result = vkQueuePresentKHR(_device->presentationQueue, &presentInfo);
-    if (result == VK_ERROR_OUT_OF_DATE_KHR || result == VK_SUBOPTIMAL_KHR
-        || this->_framebufferResized) {
-        ASSERT(
-            _tetraMode != TetraMode::kEvenOddHardwareSync
-        ); // exclusive window does not resize its swapchain
-        this->recreateSwapChain(_swapChain);
-        recreateVirtualFrameBuffers();
-        this->_framebufferResized = false;
-    }
-}
-
 // FIXME: glfw calls from a differnt thread; may need to add critical sections
 // currently for perf reasons we're leaving it as is.
 void Tetrium::bindDefaultInputs()
@@ -1985,104 +1578,4 @@ void Tetrium::bindDefaultInputs()
             }
         }
     );
-}
-
-void Tetrium::getMainProjectionMatrix(glm::mat4& projectionMatrix)
-{
-    auto& extent = _swapChain.extent;
-    projectionMatrix = glm::perspective(
-        glm::radians(_FOV),
-        extent.width / static_cast<float>(extent.height),
-        DEFAULTS::ZNEAR,
-        DEFAULTS::ZFAR
-    );
-    projectionMatrix[1][1] *= -1; // invert for vulkan coord system
-}
-
-// whether to use the new virtual frame counter
-// the old counter uses a nanosecond-precision timer,
-// the new counter takes the max of the image id so far presented.
-// TODO: profile precision of the new counter vs. old counter
-#define NEW_VIRTUAL_FRAMECOUNTER 1
-
-uint64_t Tetrium::getSurfaceCounterValue()
-{
-    uint64_t surfaceCounter;
-    switch (_tetraMode) {
-    case TetraMode::kEvenOddSoftwareSync: {
-        uint32_t imageCount;
-#if NEW_VIRTUAL_FRAMECOUNTER
-#if __APPLE__
-        vkGetPastPresentationTimingGOOGLE(
-            _device->logicalDevice, _swapChain.chain, &imageCount, nullptr
-        );
-        std::vector<VkPastPresentationTimingGOOGLE> images(imageCount);
-        vkGetPastPresentationTimingGOOGLE(
-            _device->logicalDevice, _swapChain.chain, &imageCount, images.data()
-        );
-        auto& ctx = _softwareEvenOddCtx;
-        for (int i = 0; i < imageCount; i++) {
-            auto& img = images.at(i);
-            if (img.presentID > ctx.lastPresentedImageId) {
-                ctx.numFramesPresented += 1;
-                ctx.lastPresentedImageId = img.presentID;
-            }
-            _softwareEvenOddCtx.mostRecentPresentFinish
-                = std::max(_softwareEvenOddCtx.mostRecentPresentFinish, img.actualPresentTime);
-            // INFO(
-            //     "i = {}; present id: {} : expected: {} actual: {}",
-            //     i,
-            //     img.presentID,
-            //     img.desiredPresentTime,
-            //     img.actualPresentTime
-            // );
-        }
-        surfaceCounter = ctx.numFramesPresented;
-#else  // !__APPLE__
-        NEEDS_IMPLEMENTATION(); // only apple device supports software counter so far.
-#endif // __APPLE__
-#else  // ! NEW_VIRTUAL_FRAMECOUNTER
-       // old method: count the time
-       // return a software-based surface counter
-        unsigned long long timeSinceStartNanoSeconds
-            = std::chrono::duration<double, std::chrono::nanoseconds::period>(
-                  std::chrono::steady_clock().now() - _softwareEvenOddCtx.timeEngineStart
-              )
-                  .count()
-              + _softwareEvenOddCtx.timeOffset;
-        surfaceCounter = timeSinceStartNanoSeconds / _softwareEvenOddCtx.nanoSecondsPerFrame;
-#endif // NEW_VIRTUAL_FRAMECOUNTER
-    } break;
-    case TetraMode::kEvenOddHardwareSync:
-#if WIN32
-        NEEDS_IMPLEMENTATION();
-#endif // WIN32
-#if __APPLE__
-        NEEDS_IMPLEMENTATION();
-#endif
-#if __linux__
-        _hardWareEvenOddCtx.vkGetSwapchainCounterEXT(
-            _device->logicalDevice,
-            _swapChain.chain,
-            VkSurfaceCounterFlagBitsEXT::VK_SURFACE_COUNTER_VBLANK_EXT,
-            &surfaceCounter
-        );
-#endif
-        break;
-    default:
-        surfaceCounter = 0;
-    }
-
-    return surfaceCounter;
-}
-
-bool Tetrium::isEvenFrame()
-{
-    bool isEven = getSurfaceCounterValue() % 2 == 0;
-
-    if (_flipEvenOdd) {
-        isEven = !isEven;
-    }
-
-    return isEven;
 }
